TicketApi/src/main/java/com/tbma/book/ticket/app/controller/TicketController.java
package com.tbma.book.ticket.app.controller;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.tbma.book.ticket.app.entities.Ticket;
import com.tbma.book.ticket.app.service.TicketService;

@RestController
@RequestMapping(value = "api/tickets")
public class TicketController {

	@Autowired
	TicketService ticketService;

	@PostMapping(value = "/create")
	public Ticket createTicket(@RequestBody Ticket ticket) {
		return ticketService.createTicket(ticket);
	}

	@GetMapping(value = "/ticket/{ticketId}")
	public Ticket getTicketById(@PathVariable("ticketId") Integer ticketId) {
		return ticketService.getTicketById(ticketId);
	}

	@GetMapping(value = "/ticket/alltickets")
	public Iterable<Ticket> getAllTickets() {
		return ticketService.getAllTickets();
	}

	@DeleteMapping(value = "/ticket/{ticketId}")
	public void deleteTicket(@PathVariable("ticketId") Integer ticketId) {
		ticketService.deleteTicket(ticketId);
	}

	@PutMapping(value = "/ticket/{ticketId}/{newEmail:.+}")
	public Ticket updateTicket(@PathVariable("ticketId") int ticketId, @PathVariable("newEmail") String email) {
		return ticketService.updateTicket(ticketId, email);
	}
}
